Refactor ok
