+++ src/App.js
@@ -1,8 +1,8 @@
 import { Header } from "./components/_layouts/Header";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { Footer } from "./components/_layouts/Footer";
-import { Spanish } from "./components/spanish/Spanish";
-import { English } from "./components/english/English";
+import { Spanish } from "./pages/spanish/Spanish";
+import { English } from "./pages/english/English";
 import { ErrorNotFound } from "./components/error/Error";
 
 function App() {
