+++ src/utils/utils.language.js
@@ -6,3 +6,8 @@ export const Languages = {
     node: 'Node',
     api: 'API'
 }
+
+export const Idiom = {
+    english: 'english',
+    spanish: 'spanish'
+}
